backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appnope==0.1.4
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
av==15.1.0
backcall==0.2.0
beautifulsoup4==4.14.3
bleach==6.2.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.1.8
coloredlogs==15.0.1
ctranslate2==4.7.1
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
exceptiongroup==1.3.1
executing==2.2.1
faster-whisper==1.2.1
fastjsonschema==2.21.2
filelock==3.19.1
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2025.10.0
geographiclib==2.1
geopy==2.4.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
humanfriendly==10.0
idna==3.12
importlib_metadata==8.7.1
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.84
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
markdown-it-py==3.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.17.1
nbformat==5.10.4
numpy==2.0.2
ollama==0.6.1
onnxruntime==1.19.2
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandocfilters==1.5.1
parso==0.8.6
pathvalidate==3.3.1
pexpect==4.9.0
pickleshare==0.7.5
piper-tts==1.4.2
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.6
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycparser==2.23
pydantic==2.13.3
pydantic-settings==2.11.0
pydantic_core==2.46.3
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==15.0.0
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
tinycss2==1.4.0
tokenizers==0.22.2
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.2
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
wcwidth==0.6.0
webencodings==0.5.1
xxhash==3.6.0
yarg==0.1.9
yarl==1.22.0
zipp==3.23.1
zstandard==0.25.0
